feat: 효과음을 재생하는 util 함수 구현
